Intoduce new ILongCurve to handle higher precision curve data. This effects * Cancellation fees * Lumpsum cooldown costs * Vessel charter rates
Bugzid: 1992

+++ com-mmxlabs-models-lng-transformer/com.mmxlabs.models.lng.transformer.extensions/src/com/mmxlabs/models/lng/transformer/extensions/redirection/IVoyageCostCalculator.java
@@ -15,11 +15,10 @@ import com.mmxlabs.scheduler.optimiser.voyage.impl.VoyagePlan;
 public interface IVoyageCostCalculator {
 
 	@Nullable
-	VoyagePlan calculateShippingCosts(@NonNull IPort loadPort, @NonNull IPort dischargePort, int loadTime, int loadDuration, int dischargeTime, int dischargeDuration,
-			@NonNull IVessel vessel, int vesselCharterInRatePerDay, long startHeelInM3, int notionalBallastSpeed, int cargoCVValue, @NonNull ERouteOption route,
-			int baseFuelPricePerMT, int salesPricePerMMBTu);
+	VoyagePlan calculateShippingCosts(@NonNull IPort loadPort, @NonNull IPort dischargePort, int loadTime, int loadDuration, int dischargeTime, int dischargeDuration, @NonNull IVessel vessel,
+			long vesselCharterInRatePerDay, long startHeelInM3, int notionalBallastSpeed, int cargoCVValue, @NonNull ERouteOption route, int baseFuelPricePerMT, int salesPricePerMMBTu);
 
 	@Nullable
 	VoyagePlan calculateShippingCosts(@NonNull IPort loadPort, @NonNull IPort dischargePort, int loadTime, int loadDuration, int dischargeTime, int dischargeDuration, int returnTime,
-			@NonNull IVessel vessel, int vesselCharterInRatePerDay, long startHeelInM3, int cargoCVValue, @NonNull ERouteOption route, int baseFuelPricePerMT, int salesPricePerMMBTu);
+			@NonNull IVessel vessel, long vesselCharterInRatePerDay, long startHeelInM3, int cargoCVValue, @NonNull ERouteOption route, int baseFuelPricePerMT, int salesPricePerMMBTu);
 }
